# docs/design/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Team Document Internal API
  version: 1.0.0
  description: Browser-facing internal contract. Not a customer Public API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /workspaces/{workspaceId}/documents/{documentId}/backlinks:
    get:
      tags: [Knowledge]
      operationId: listBacklinks
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {description: Backlinks, content: {application/json: {schema: {$ref: '#/components/schemas/ReferencePage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/references:
    post:
      tags: [Knowledge]
      operationId: createReference
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/LeaseToken'}, {$ref: '#/components/parameters/ClientInstance'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/CreateReference'}
      responses:
        '201': {description: Reference, content: {application/json: {schema: {$ref: '#/components/schemas/Reference'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/references/{referenceId}:
    delete:
      tags: [Knowledge]
      operationId: deleteReference
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/ReferenceId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/LeaseToken'}, {$ref: '#/components/parameters/ClientInstance'}, {$ref: '#/components/parameters/CsrfToken'}]
      responses:
        '204': {description: Deleted}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/vocabulary:
    get:
      tags: [Knowledge]
      operationId: listVocabulary
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {description: Vocabulary, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyPage'}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Knowledge]
      operationId: createVocabularyConcept
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/WriteVocabularyConcept'}
      responses:
        '201': {description: Concept, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyConcept'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/vocabulary/{conceptId}:
    get:
      tags: [Knowledge]
      operationId: getVocabularyConcept
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/ConceptId'}]
      responses:
        '200': {description: Concept, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyConcept'}}}}
        default: {$ref: '#/components/responses/Problem'}
    put:
      tags: [Knowledge]
      operationId: updateVocabularyConcept
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/ConceptId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/WriteVocabularyConcept'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyConcept'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/vocabulary/{conceptId}/deprecate:
    post:
      tags: [Knowledge]
      operationId: deprecateVocabularyConcept
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/ConceptId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/DeprecateVocabularyConcept'}
      responses:
        '200': {description: Deprecated, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyConcept'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: adoc_session}
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
    DocumentId: {name: documentId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
...
    ReferenceId: {name: referenceId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
    ConceptId: {name: conceptId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 2048}}
...
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string, pattern: '^"[0-9]+"$'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    CsrfToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 43, maxLength: 1024}}
    LeaseToken: {name: X-Edit-Lease, in: header, required: true, schema: {type: string}}
    ClientInstance: {name: X-Client-Instance, in: header, required: true, schema: {$ref: '#/components/schemas/Id'}}
  requestBodies:
...
    CreateReference:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [sourceRegion, target], properties: {sourceRegion: {$ref: '../contracts/document-operation.schema.json#/$defs/region'}, target: {$ref: '../contracts/document-operation.schema.json#/$defs/referenceTarget'}}}}}
    WriteVocabularyConcept:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [canonicalTerm, definition, terms], properties: {canonicalTerm: {type: string, minLength: 1, maxLength: 200}, definition: {type: string, minLength: 1, maxLength: 5000}, terms: {type: array, minItems: 1, maxItems: 100, items: {$ref: '#/components/schemas/VocabularyTerm'}}}}}}
    DeprecateVocabularyConcept:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [reason], properties: {reason: {type: string, minLength: 1, maxLength: 1000}, replacementConceptId: {$ref: '#/components/schemas/NullableId'}}}}}
...
  responses:
    Problem:
      description: Domain or transport problem
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
    Id: {type: string, format: uuid}
    NullableId: {oneOf: [{$ref: '#/components/schemas/Id'}, {type: 'null'}]}
...
    Reference: {type: object, additionalProperties: false, required: [id, sourceDocumentId, sourceRegion, target, snapshot, createdAt], properties: {id: {$ref: '#/components/schemas/Id'}, sourceDocumentId: {$ref: '#/components/schemas/Id'}, sourceRegion: {$ref: '../contracts/document-operation.schema.json#/$defs/region'}, target: {$ref: '../contracts/document-operation.schema.json#/$defs/referenceTarget'}, snapshot: {type: object, additionalProperties: false, required: [title, snapshotHash], properties: {title: {type: string}, snapshotHash: {type: string, pattern: '^[a-f0-9]{64}$'}}}, createdAt: {type: string, format: date-time}}}
    ReferencePage: {type: object, additionalProperties: false, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Reference'}}, nextCursor: {type: [string, 'null']}}}
    VocabularyTerm: {type: object, additionalProperties: false, required: [term, kind], properties: {term: {type: string, minLength: 1, maxLength: 200}, kind: {type: string, enum: [CANONICAL, SYNONYM, PROHIBITED]}}}
    VocabularyConcept: {type: object, additionalProperties: false, required: [id, canonicalTerm, definition, terms, status, replacementConceptId, revision], properties: {id: {$ref: '#/components/schemas/Id'}, canonicalTerm: {type: string}, definition: {type: string}, terms: {type: array, items: {$ref: '#/components/schemas/VocabularyTerm'}}, status: {type: string, enum: [ACTIVE, DEPRECATED]}, replacementConceptId: {$ref: '#/components/schemas/NullableId'}, revision: {type: integer, minimum: 0}}}
    VocabularyPage: {type: object, additionalProperties: false, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/VocabularyConcept'}}, nextCursor: {type: [string, 'null']}}}
...
    Problem: {type: object, required: [type, title, status, code, retryable, correlationId], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, code: {type: string}, retryable: {type: boolean}, correlationId: {type: string}, currentRevision: {type: integer}, baseVersionId: {$ref: '#/components/schemas/NullableId'}, currentVersionId: {$ref: '#/components/schemas/NullableId'}, draftId: {$ref: '#/components/schemas/NullableId'}, fieldErrors: {type: array, items: {type: object, required: [field, code], properties: {field: {type: string}, code: {type: string}}}}}}